IDP-1972 initializing commonConfirmTrx Api
