+++ src/main/java/it/gov/pagopa/payment/controller/payment/CommonPaymentController.java
@@ -2,6 +2,7 @@ package it.gov.pagopa.payment.controller.payment;
 
 import it.gov.pagopa.payment.dto.common.BaseTransactionResponseDTO;
 import it.gov.pagopa.payment.dto.qrcode.TransactionCreationRequest;
+import it.gov.pagopa.payment.dto.qrcode.TransactionResponse;
 import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
@@ -14,4 +15,10 @@ public interface CommonPaymentController {
             @RequestHeader("x-merchant-id") String merchantId,
             @RequestHeader("x-acquirer-id") String acquirerId,
             @RequestHeader("x-apim-request-id") String idTrxIssuer);
+
+    @PutMapping("/{transactionId}/confirm")
+    TransactionResponse confirmPayment(
+            @PathVariable("transactionId") String trxId,
+            @RequestHeader("x-merchant-id") String merchantId,
+            @RequestHeader("x-acquirer-id") String acquirerId);
 }
